+++ util.h
@@ -24,6 +24,8 @@
  * SUCH DAMAGE.
  */
 
+#include <stdint.h>
+
 extern int gDebug;
 
 #if defined(DEBUG)
@@ -36,7 +38,7 @@ extern int gDebug;
 #define DEBUG_TUNNEL	(0)
 #endif
 
-extern char *iptoa (u_long ip);
-extern u_long atoip (char *s);
+extern char *iptoa (uint32_t ip);
+extern uint32_t atoip (char *s);
